tests/wpt/web-platform-tests/css/css-ui/parsing/user-select-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: getComputedValue().userSelect</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#content-selection">
<meta name="assert" content="user-select computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("user-select", "auto");
test_computed_value("user-select", "text");
test_computed_value("user-select", "none");
test_computed_value("user-select", "contain");
test_computed_value("user-select", "all");
</script>
</body>
</html>
