LayoutTests/imported/w3c/web-platform-tests/css/css-ui/parsing/user-select-computed-expected.txt


FAIL Property user-select value 'auto' assert_equals: expected "auto" but got "text"
PASS Property user-select value 'text'
PASS Property user-select value 'none'
FAIL Property user-select value 'contain' assert_true: 'contain' is a supported value for user-select. expected true got false
PASS Property user-select value 'all'
